requirements.txt
# Package Management
pip==24.0
setuptools==69.1.1
wheel==0.42.0

# Web Frameworks
streamlit==1.29.0
Flask==3.0.2
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.1

# Data Processing
pandas==2.0.3
numpy==1.26.3

# HTTP Requests
requests==2.28.2

# Development Tools
black==24.2.0
flake8==7.0.0

# Path handling
pathlib==1.0.1

# Environment Variables
python-dotenv==1.0.0

# Testing
pytest==8.1.1
pytest-cov==4.1.0

# Development Tools
mypy==1.8.0
isort==5.13.2

# Documentation
Sphinx==7.2.6
sphinx-rtd-theme==2.0.0

# Build and Distribution
build==1.0.3
twine==5.0.0

# Plotting
plotly==5.18.0
uuid>=1.30

# HTTP Client
aiohttp>=3.8.0

# Asyncio
nest_asyncio>=1.5.6

# PDF生成
fpdf2>=2.7.0
reportlab>=4.0.7

# API调用
anthropic>=0.3.11

# 日志和工具
python-dotenv>=1.0.0

# 定时任务
schedule>=1.2.0

# 文件大小格式化
humanize>=4.7.0

# Word文档处理
python-docx==0.8.11

# 字体处理
fonttools>=4.0.0

# AWS相关依赖
boto3>=1.26.0
botocore>=1.29.0